Consolidate indexer accuracy research docs and reorganize tests
- Replace round-by-round accuracy research notes with a single summary
  doc
- Update references in guides and research links
- Rewrite indexer README for DeGov context and add developer docs links
- Move test files into unit, integration, and accuracy subfolders
- Update test imports for new structure
- Add and clarify package scripts for dev, test, and audit tasks
