+++ packages/indexer/README.md
@@ -1,97 +1,81 @@
-[![Open in Gitpod](https://gitpod.io/button/open-in-gitpod.svg)](https://gitpod.io/#https://github.com/subsquid/squid-evm-template)
+# DeGov Indexer
 
-# Minimal EVM squid
+`packages/indexer` is the Subsquid-based governance indexer used by DeGov.
 
-This is a starter template of a squid indexer for EVM networks (Ethereum, Polygon, BSC, etc.). See [Squid SDK docs](https://docs.subsquid.io/) for a complete reference.
-
-To extract EVM logs and transactions by a topic or a contract address, use [`.addLog()`](https://docs.subsquid.io/evm-indexing/configuration/evm-logs/), [`.addTransaction()`](https://docs.subsquid.io/evm-indexing/configuration/transactions/), [`.addTrace()`](https://docs.subsquid.io/evm-indexing/configuration/traces/) or [`.addStateDiff()`](https://docs.subsquid.io/evm-indexing/configuration/state-diffs/) methods of the `EvmBatchProcessor` instance defined in `src/processor.ts`. Select data fields with [`.setFields()`](https://docs.subsquid.io/evm-indexing/configuration/data-selection/).
-
-The requested data is transformed in batches by a single handler provided to the `processor.run()` method. 
-
-For a full list of supported networks and config options,
-check the [`EvmBatchProcessor` overview](https://docs.subsquid.io/evm-indexing/evm-processor/) and the [setup section](https://docs.subsquid.io/evm-indexing/configuration/).
-
-For a step-by-step migration guide from TheGraph, see [the dedicated docs page](https://docs.subsquid.io/migrate/migrate-subgraph/).
-
-Dependencies: Node.js v16 or newer, Git, Docker.
+It reads DAO config, indexes governor/token/timelock events into PostgreSQL,
+serves GraphQL, and includes tooling for replay, reconciliation, audit, and
+single-address diagnosis.
 
 ## Quickstart
 
 ```bash
-# 0. Install @subsquid/cli a.k.a. the sqd command globally
-npm i -g @subsquid/cli
-
-# 1. Retrieve the template
-sqd init my_squid_name -t evm
-cd my_squid_name
-
-# 2. Install dependencies
-npm i
-
-# 3. Start a Postgres database container and detach
-sqd up
-
-# 4. Build the squid
-sqd build
-
-# 5. Start both the squid processor and the GraphQL server
-sqd run .
-```
-A GraphiQL playground will be available at [localhost:4350/graphql](http://localhost:4350/graphql).
-
-You can also start squid services one by one:
-```bash
-sqd process
-sqd serve
+cd packages/indexer
+just install
+just up
+just codegen
+just build
+just run
 ```
 
-## Dev flow
+GraphQL is served on `http://localhost:4350/graphql`.
 
-### 1. Define database schema
+## Command groups
 
-Start development by defining the schema of the target database via `schema.graphql`.
-Schema definition consists of regular graphql type declarations annotated with custom directives.
-Full description of `schema.graphql` dialect is available [here](https://docs.subsquid.io/store/postgres/schema-file/).
-
-### 2. Generate TypeORM classes
-
-Mapping developers use TypeORM [EntityManager](https://typeorm.io/#/working-with-entity-manager)
-to interact with target database during data processing. All necessary entity classes are
-generated by the squid framework from `schema.graphql`. This is done by running `sqd codegen`
-command.
-
-### 3. Generate database migrations
-
-All database changes are applied through migration files located at `db/migrations`.
-`squid-typeorm-migration(1)` tool provides several commands to drive the process.
+### Codegen
 
 ```bash
-## drop create the database
-sqd down
-sqd up
-
-## replace any old schemas with a new one made from the entities
-sqd migration:generate
+yarn run codegen:abi
+yarn run codegen:schema
+yarn run codegen
 ```
-See [docs on database migrations](https://docs.subsquid.io/store/postgres/db-migrations/) for more details.
 
-### 4. Import ABI contract and generate interfaces to decode events
-
-It is necessary to import the respective ABI definition to decode EVM logs. One way to generate a type-safe facade class to decode EVM logs is by placing the relevant JSON ABIs to `./abi`, then using `squid-evm-typegen(1)` via an `sqd` script:
+### Database
 
 ```bash
-sqd typegen
+yarn run db:migrate
+yarn run db:migrate:force
 ```
 
-See more details on the [`squid-evm-typegen` doc page](https://docs.subsquid.io/evm-indexing/squid-evm-typegen).
+### Runtime
 
-## Project conventions
+```bash
+yarn run build
+yarn run dev:start
+yarn run dev:smart-start
+yarn run dev:smart-start:force
+yarn run dev:graphql
+```
 
-Squid tools assume a certain [project layout](https://docs.subsquid.io/basics/squid-structure):
+### Tests
 
-* All compiled js files must reside in `lib` and all TypeScript sources in `src`.
-The layout of `lib` must reflect `src`.
-* All TypeORM classes must be exported by `src/model/index.ts` (`lib/model` module).
-* Database schema must be defined in `schema.graphql`.
-* Database migrations must reside in `db/migrations` and must be plain js files.
-* `sqd(1)` and `squid-*(1)` executables consult `.env` file for environment variables.
+```bash
+yarn run test:unit
+yarn run test:accuracy
+yarn run test:integration
+```
+
+### Audit
+
+```bash
+yarn run audit:accuracy
+yarn run audit:diagnose -- --address 0x... --code ens-dao
+```
+
+## Preferred developer entrypoints
+
+For day-to-day local work, prefer the package `justfile`:
+
+```bash
+just build
+just db-migrate
+just smart-start
+just test-accuracy
+just diagnose-address 0x983110309620d911731ac0932219af06091b6744 ens-dao
+```
+
+## Docs
+
+- [Indexer developer guide](../../docs/guides/20260325__indexer_developer_guide.md)
+- [Indexer accuracy diagnosis](../../docs/guides/20260331__indexer_accuracy_diagnosis.md)
+- [Indexer accuracy research](../../docs/research/20260401__indexer_accuracy_research.md)
+- [Indexer architecture](../../docs/architecture/20260325__indexer_architecture.md)
