+++ packages/indexer/package.json
@@ -7,14 +7,25 @@
     "codegen:abi": "squid-evm-typegen src/abi ./abi/*.json",
     "codegen:schema": "squid-typeorm-codegen",
     "codegen": "sqd codegen",
-    "migrate:db": "zx scripts/sqd-migration.mjs",
+    "db:migrate": "zx scripts/sqd-migration.mjs",
+    "db:migrate:force": "zx scripts/sqd-migration.mjs --force",
+    "migrate:db": "yarn run db:migrate",
     "build": "sqd build",
-    "reconcile": "node lib/reconcile.js",
-    "replay:backfill": "./scripts/replay-backfill.sh",
-    "test": "jest --runInBand",
-    "test:integration": "jest --runInBand --config jest.integration.config.js --runTestsByPath __tests__/chaintool.integration.test.ts __tests__/textplus.integration.test.ts",
+    "dev:start": "sh ./scripts/start.sh",
+    "dev:smart-start": "sh ./scripts/smart-start.sh",
+    "dev:smart-start:force": "sh ./scripts/smart-start.sh force",
+    "dev:graphql": "sh ./scripts/graphql-server.sh",
+    "dev:reconcile": "node lib/reconcile.js",
+    "dev:replay-backfill": "./scripts/replay-backfill.sh",
+    "reconcile": "yarn run dev:reconcile",
+    "replay:backfill": "yarn run dev:replay-backfill",
+    "test": "yarn run test:unit",
+    "test:unit": "jest --runInBand",
+    "test:accuracy": "jest --runInBand __tests__/accuracy/indexerAccuracyAudit.test.ts __tests__/accuracy/indexerAccuracyDiagnose.test.ts __tests__/accuracy/token-vote-power.test.ts",
+    "test:integration": "jest --runInBand --config jest.integration.config.js --runTestsByPath __tests__/integration/chaintool.integration.test.ts __tests__/integration/textplus.integration.test.ts",
     "audit:accuracy": "node scripts/indexer-accuracy-audit.js",
-    "audit:diagnose-address": "node scripts/indexer-accuracy-diagnose.js"
+    "audit:diagnose": "node scripts/indexer-accuracy-diagnose.js",
+    "audit:diagnose-address": "yarn run audit:diagnose"
   },
   "dependencies": {
     "@openrouter/ai-sdk-provider": "^2.3.3",
